server/requirements.txt
flask==3.0.3 
sqlalchemy==2.0.30
pg8000==1.31.2 
python-dotenv==1.0.1 
firebase-admin==6.5.0 
flask-session==0.8.0 
cloud-sql-python-connector==1.9.2
gunicorn==22.0.0
flask-socketio==5.3.6
websocket-client==1.8.0  

google-api-core==2.19.0 
google-api-python-client==2.129.0 
google-auth==2.29.0 
google-auth-httplib2==0.2.0 
google-cloud-core==2.4.1 
google-cloud-firestore==2.16.0 
google-cloud-storage==2.16.0 
google-crc32c==1.5.0 
google-resumable-media==2.7.0 
googleapis-common-protos==1.63.0 
